Move search component to separate folder
