+++ react/src/components/index.tsx
@@ -14,7 +14,7 @@ import ErrorProvider, {
 } from './Error/ErrorProvider';
 import ErrorFallback from './ErrorFallback';
 import ErrorIndicator from './ErrorIndicator';
-import GeneNameSearch from './GeneNameSearch';
+import GeneNameSearch from './Search/GeneNameSearch';
 import IconButton from './IconButton';
 import Input from './Input';
 import { Background, Body, ButtonWrapper, Column, Container, Flex, InlineFlex } from './Layout';
